+++ feature/src/main/java/com/zcard/feature/cardeditor/CardEditorViewModel.kt
@@ -26,8 +26,8 @@ import com.zcard.domain.repository.AssetRepository
 import com.zcard.domain.repository.CardElementRepository
 import com.zcard.domain.repository.CardRepository
 import com.zcard.domain.usecase.GenerateCardUrlUseCase
+import com.zcard.domain.usecase.GetCardUseCase
 import com.zcard.domain.usecase.GetTextElementsUseCase
-import com.zcard.domain.usecase.OpenCardEditorUseCase
 import com.zcard.domain.usecase.SaveTextElementsParams
 import com.zcard.domain.usecase.SaveTextElementsUseCase
 import com.zcard.domain.usecase.UploadCardModelUseCase
@@ -57,7 +57,7 @@ class CardEditorViewModel @Inject constructor(
     private val cardRepository: CardRepository,
     private val assetRepository: AssetRepository,
     private val cardElementRepository: CardElementRepository,
-    private val openCardEditorUseCase: OpenCardEditorUseCase,
+    private val getCardUseCase: GetCardUseCase,
     private val getTextsUseCase: GetTextElementsUseCase,
     private val saveTextElementsUseCase: SaveTextElementsUseCase,
     private val generateCardUrlUseCase: GenerateCardUrlUseCase,
@@ -142,7 +142,7 @@ class CardEditorViewModel @Inject constructor(
 
     private fun handleInit(cardId: Long) {
         viewModelScope.launch {
-            openCardEditorUseCase(cardId)
+            getCardUseCase(cardId)
                 .onSuccess { result ->
                     _cardId = result.cardData.cardId
                     _exportId = result.cardData.exportId
